application/storages/databases/migrations/models/0_empty.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        SELECT 1;
        """


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        SELECT 1;
        """
