pyproject.toml
[tool.poetry]
name = "sbersupermarket-parser"
version = "0.0.1"
description = "Fast-api application for getting data from sbersupermarket"
authors = ["Alexandr Bogut"]
readme = "README.md"
license = "MIT"
packages = [
    { include = "application" },
]


[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.21.1"
fastapi = "^0.95.1"
pydantic = "^1.7.3"
loguru = "^0.6.0"
sentry-sdk = "^1.12.1"
undetected-chromedriver = "^3.4.7"
py-memoize = "^1.1.3"
aiomisc = "^17.1.2"
retry = "^0.9.2"
starlette-context = "^0.3.6"
pytz = "^2023.3"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
sqlalchemy = "^1.3.24"
asyncpg = "^0.27.0"
aerich = "^0.7.1"


[tool.poetry.group.dev.dependencies]
black = {version = "^18.3-alpha.0", allow-prereleases = true}
pre-commit = "^2.9.3"
mypy = "^0.991"
requests = "^2.25.0"
pytest = "^6.2.0"
pytest-asyncio = "^0.14.0"
pytest-cov = "^2.10.1"
mkdocs = "^1.1.2"
mkdocs-material = "^4.6.3"
httpx = "^0.23.1"
types-retry = "^0.9.9.3"
types-pytz = "^2023.3.0.0"
vulture = "^2.7"
python-dotenv = "^1.0.0"
mock = "^5.0.2"


[tool.aerich]
tortoise_orm = "application.storages.databases.connector.DEFAULT_TORTOISE_ORM_CONFIG"
location = "./application/storages/databases/migrations"
src_folder = "./."
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
